+++ Assets/Scripts/ClickOptions/ClickUITextDetector.cs
@@ -5,11 +5,13 @@ using UnityEngine.UI;
 public class ClickUITextDetector : MonoBehaviour, IPointerClickHandler
 {
 	private Text text;
+	private QuestionOne q;
 	public GameObject textGameObject;
-    public bool IsAnswerCorrect { get; private set; }
+	public bool IsAnswerCorrect { get; private set; }
 
-    void Start()
+	void Start()
 	{
+ q = GetComponentInParent<QuestionOne>();
 		text = textGameObject.GetComponent<Text>();
 	}
 
@@ -34,32 +36,33 @@ public class ClickUITextDetector : MonoBehaviour, IPointerClickHandler
 			{
 				string questionName = questionObject.name;
 
-				IsAnswerCorrect = Judege(option, questionName);
+				IsAnswerCorrect = q.Judege(option, questionName);
+
+				Debug.Log("点击的选项所属问题是: " + IsAnswerCorrect);
 
 				textGameObject.SetActive(true);
 				text.text = option;
-				Debug.Log("点击的选项所属问题是: " + IsAnswerCorrect);
 			}
 		}
 	}
-
-	private string[] questions = { "QuestionOne", "QuestionTwo", "QuestionThree", "QuestionFour", "QuestionFive" };
-
-	private string[] options = {"A" , "B" , "A" , "C" , "B"};
-
-	public bool Judege(string option , string questionName)
-	{
-		int count = 0;
-
-		for (int i = 0; i < questions.Length; i++)
-		{
-			if (questions[i] == questionName)
-			{
-				count = i;
-				break;
-			}
-		}
-
-		return options[count] == option.Split(' ')[0];
-	}
+	//
+	// private string[] questions = { "QuestionOne", "QuestionTwo", "QuestionThree", "QuestionFour", "QuestionFive" };
+	//
+	// private string[] options = {"A" , "B" , "A" , "C" , "B"};
+	//
+	// public bool Judege(string option , string questionName)
+	// {
+	// 	int count = 0;
+	//
+	// 	for (int i = 0; i < questions.Length; i++)
+	// 	{
+	// 		if (questions[i] == questionName)
+	// 		{
+	// 			count = i;
+	// 			break;
+	// 		}
+	// 	}
+	//
+	// 	return options[count] == option.Split(' ')[0];
+	// }
 }
