+++ Assets/Scripts/NPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 553caf90e7051fa49adc572e6ab52fcf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
