+++ Assets/Scripts/PlayerController/PlayerController.cs
@@ -1,17 +1,13 @@
-using UnityEngine; using UnityEngine.UI;
+using UnityEngine;
 
 public class PlayerController : MonoBehaviour
 {
 	public float nextLevelPosition;
 	public float preLevelPosition;
-	public Image mainPlane;
-	public Image dialogueButton;
+
 	public float moveSpeed = 1f;
 	public Rigidbody2D player;
 	private Vector2 movement;
-
-	private bool isDialogue;
-	private bool toggleDialoguePlane;
 
 	NextLevelScence nextLevelScence = new NextLevelScence();
 
@@ -22,41 +18,13 @@ public class PlayerController : MonoBehaviour
 	void Update()
 	{
 		Walk();
-		Dialogue();
 		NextScene();
-	}
-
-	void Dialogue(){
-		if (Input.GetKeyDown(KeyCode.E) && isDialogue)
-		{
-			mainPlane.gameObject.SetActive(!toggleDialoguePlane);
-			dialogueButton.gameObject.SetActive(toggleDialoguePlane);
-			toggleDialoguePlane = !toggleDialoguePlane;
-		}
 	}
 
 	void Walk(){
 		float moveDir = Input.GetAxis("Horizontal");
 		Vector2 movement = new Vector2(moveDir, player.velocity.y);
 		player.MovePosition(player.position + movement * moveSpeed * Time.fixedDeltaTime);
-	}
-
-void OnTriggerEnter2D(Collider2D other)
-	{
-		if (other.CompareTag("NPC"))
-		{
-			dialogueButton.gameObject.SetActive(true);
-			isDialogue = true;
-		}
-	}
-
-	void OnTriggerExit2D(Collider2D other)
-	{
-		if (other.CompareTag("NPC"))
-		{
-			dialogueButton.gameObject.SetActive(false);
-			isDialogue = false;
-		}
 	}
 
 	void NextScene(){
