+++ UserSettings/EditorUserSettings.asset
@@ -9,16 +9,16 @@ EditorUserSettings:
       value: 50010302030d5f590e5f097144720c4445161b7c2f7f2766742b4430b6e3603d
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 5153500404045c0b5e5d5b2643220944144f4b2c2e2d71617f7f4d6ae4b06268
-      flags: 0
-    RecentlyUsedSceneGuid-2:
       value: 0105505f50570b035f0f5c23422209441315417d2a2c2033787b4e67b5b8363c
       flags: 0
-    RecentlyUsedSceneGuid-3:
+    RecentlyUsedSceneGuid-2:
       value: 52050600000d0b0c580b09274176084415154d7a2a7a75627a784d31e0b5316c
       flags: 0
-    RecentlyUsedSceneGuid-4:
+    RecentlyUsedSceneGuid-3:
       value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 5153500404045c0b5e5d5b2643220944144f4b2c2e2d71617f7f4d6ae4b06268
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
